Add Content IDs to endpoint to create Groups
A group is an association of content_ids, and as we have no further 
requirements of what we are going to do with them, we are just storing
the content_ids as an array within the Group.

This approach has drawbacks in the long run if we decide to add many
logic to them, but the approach is so simple to change, that it seems
a better approach to keep things simple, and not make further modelling
decisions until we have a clear picture of what we are going to do with
them.

I am not creating a many-to-many association with ContentItem as the 
ContentItem is something that will stop existing as a persisted model
within the Content Performance Manager.

I have done some tests persisting long groups of ContentIDs in a group
(around 10_000), and the query takes a bit longer (as expected), but still
the performance is good enough to deploy it.