Add parent group to Endpoint to create Groups
Groups follow a hierarchical nature, so we should be able to assign
a parent group to any parent. We have decided to use the Group slug
to assign the parent, because it seems easier to remember to the API 
user than the Group ID, which is a number that is internal to the
database.